# ast: avoid extra work parsing objects vs obj comprehensions