@srenatus 2021-08-19
Good idea to check the cost here. This is the benchstat of all "BenchmarkParse*" benchmarks:

```
name                                 old time/op    new time/op    delta
ParseModuleRulesBase/1-16              4.13µs ± 2%    5.91µs ±21%   +43.15%  (p=0.008 n=5+5)
ParseModuleRulesBase/10-16             30.1µs ± 2%    41.9µs ± 3%   +39.27%  (p=0.008 n=5+5)
ParseModuleRulesBase/100-16             296µs ± 1%     402µs ± 3%   +35.84%  (p=0.008 n=5+5)
ParseModuleRulesBase/1000-16           3.49ms ± 9%    4.64ms ± 6%   +32.87%  (p=0.008 n=5+5)
ParseStatementBasicCall-16             2.69µs ± 4%    3.67µs ± 5%   +36.27%  (p=0.008 n=5+5)
ParseStatementMixedJSON-16             78.8µs ± 8%   154.4µs ± 7%   +95.85%  (p=0.008 n=5+5)
ParseStatementSimpleArray/1-16         2.18µs ± 1%    3.03µs ± 2%   +38.82%  (p=0.008 n=5+5)
ParseStatementSimpleArray/10-16        7.34µs ± 2%   13.37µs ±10%   +82.29%  (p=0.008 n=5+5)
ParseStatementSimpleArray/100-16       58.6µs ± 3%   104.6µs ± 1%   +78.47%  (p=0.008 n=5+5)
ParseStatementSimpleArray/1000-16       580µs ± 3%    1064µs ± 3%   +83.25%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x1-16     3.12µs ± 1%    4.27µs ± 3%   +36.68%  (p=0.008 n=5+5)
ParseStatementNestedObjects/5x1-16     8.42µs ± 2%   13.36µs ± 2%   +58.78%  (p=0.008 n=5+5)
ParseStatementNestedObjects/10x1-16    14.7µs ± 1%    24.9µs ± 2%   +69.81%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x5-16     8.16µs ± 3%   13.29µs ± 3%   +62.83%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x10-16    13.8µs ± 3%    24.8µs ± 2%   +79.23%  (p=0.008 n=5+5)
ParseStatementNestedObjects/5x5-16     4.94ms ± 3%   10.16ms ± 4%  +105.51%  (p=0.008 n=5+5)
ParseBasicABACModule-16                47.5µs ± 3%    64.8µs ± 6%   +36.40%  (p=0.008 n=5+5)

name                                 old alloc/op   new alloc/op   delta
ParseModuleRulesBase/1-16              3.97kB ± 0%    6.01kB ± 0%   +51.41%  (p=0.008 n=5+5)
ParseModuleRulesBase/10-16             26.0kB ± 0%    40.2kB ± 0%   +54.33%  (p=0.008 n=5+5)
ParseModuleRulesBase/100-16             252kB ± 0%     388kB ± 0%   +53.66%  (p=0.008 n=5+5)
ParseModuleRulesBase/1000-16           2.56MB ± 0%    4.02MB ± 0%   +57.16%  (p=0.008 n=5+5)
ParseStatementBasicCall-16             3.02kB ± 0%    4.50kB ± 0%   +48.69%  (p=0.008 n=5+5)
ParseStatementMixedJSON-16             51.9kB ± 0%   136.2kB ± 0%  +162.33%  (p=0.008 n=5+5)
ParseStatementSimpleArray/1-16         2.86kB ± 0%    4.33kB ± 0%   +51.54%  (p=0.008 n=5+5)
ParseStatementSimpleArray/10-16        5.86kB ± 0%   13.03kB ± 0%  +122.58%  (p=0.008 n=5+5)
ParseStatementSimpleArray/100-16       38.4kB ± 0%   106.3kB ± 0%  +177.07%  (p=0.008 n=5+5)
ParseStatementSimpleArray/1000-16       390kB ± 0%    1125kB ± 0%  +188.65%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x1-16     3.66kB ± 0%    5.70kB ± 0%   +55.80%  (p=0.008 n=5+5)
ParseStatementNestedObjects/5x1-16     6.44kB ± 0%   13.62kB ± 0%  +111.46%  (p=0.016 n=5+4)
ParseStatementNestedObjects/10x1-16    10.0kB ± 0%    24.1kB ± 0%  +141.18%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x5-16     8.51kB ± 0%   15.69kB ± 0%   +84.34%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x10-16    14.6kB ± 0%    28.7kB ± 0%   +97.07%  (p=0.008 n=5+5)
ParseStatementNestedObjects/5x5-16     3.42MB ± 0%    9.13MB ± 0%  +167.25%  (p=0.008 n=5+5)
ParseBasicABACModule-16                35.1kB ± 0%    56.8kB ± 0%   +61.87%  (p=0.008 n=5+5)

name                                 old allocs/op  new allocs/op  delta
ParseModuleRulesBase/1-16                80.0 ± 0%      97.0 ± 0%   +21.25%  (p=0.008 n=5+5)
ParseModuleRulesBase/10-16                600 ± 0%       709 ± 0%   +18.17%  (p=0.008 n=5+5)
ParseModuleRulesBase/100-16             5.83k ± 0%     6.85k ± 0%   +17.53%  (p=0.008 n=5+5)
ParseModuleRulesBase/1000-16            58.0k ± 0%     68.1k ± 0%   +17.31%  (p=0.016 n=4+5)
ParseStatementBasicCall-16               47.0 ± 0%      59.0 ± 0%   +25.53%  (p=0.008 n=5+5)
ParseStatementMixedJSON-16              1.02k ± 0%     1.59k ± 0%   +56.74%  (p=0.008 n=5+5)
ParseStatementSimpleArray/1-16           36.0 ± 0%      48.0 ± 0%   +33.33%  (p=0.008 n=5+5)
ParseStatementSimpleArray/10-16           112 ± 0%       170 ± 0%   +51.79%  (p=0.008 n=5+5)
ParseStatementSimpleArray/100-16          837 ± 0%      1353 ± 0%   +61.65%  (p=0.008 n=5+5)
ParseStatementSimpleArray/1000-16       8.05k ± 0%    13.09k ± 0%   +62.59%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x1-16       49.0 ± 0%      66.0 ± 0%   +34.69%  (p=0.008 n=5+5)
ParseStatementNestedObjects/5x1-16        121 ± 0%       179 ± 0%   +47.93%  (p=0.008 n=5+5)
ParseStatementNestedObjects/10x1-16       207 ± 0%       316 ± 0%   +52.66%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x5-16        125 ± 0%       183 ± 0%   +46.40%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x10-16       220 ± 0%       329 ± 0%   +49.55%  (p=0.008 n=5+5)
ParseStatementNestedObjects/5x5-16      71.1k ± 0%    110.4k ± 0%   +55.22%  (p=0.008 n=5+5)
ParseBasicABACModule-16                   841 ± 0%       997 ± 0%   +18.55%  (p=0.008 n=5+5)
```
💭 Relatively bad but not absolutely catastrophic.