@srenatus 2021-08-19
So, this is where we're at now:

```
name                                     old time/op    new time/op    delta
ParseStatementNestedObjectsOrSets/1-16     3.95µs ± 1%    4.68µs ± 3%   +18.50%  (p=0.008 n=5+5)
ParseStatementNestedObjectsOrSets/5-16     79.9µs ± 1%    19.6µs ±13%   -75.44%  (p=0.008 n=5+5)
ParseStatementNestedObjectsOrSets/10-16    2.64ms ± 1%    0.04ms ± 4%   -98.65%  (p=0.008 n=5+5)
ParseStatementNestedObjectsOrSets/15-16    87.5ms ± 1%     0.1ms ± 2%   -99.94%  (p=0.008 n=5+5)
ParseStatementNestedObjectsOrSets/20-16     2.88s ± 3%     0.00s ± 6%  -100.00%  (p=0.008 n=5+5)

name                                     old alloc/op   new alloc/op   delta
ParseStatementNestedObjectsOrSets/1-16     3.97kB ± 0%    5.07kB ± 0%   +27.63%  (p=0.008 n=5+5)
ParseStatementNestedObjectsOrSets/5-16     56.8kB ± 0%    16.0kB ± 0%   -71.74%  (p=0.000 n=4+5)
ParseStatementNestedObjectsOrSets/10-16    1.86MB ± 0%    0.03MB ± 0%   -98.38%  (p=0.008 n=5+5)
ParseStatementNestedObjectsOrSets/15-16    62.9MB ± 0%     0.0MB ± 0%   -99.93%  (p=0.008 n=5+5)
ParseStatementNestedObjectsOrSets/20-16    2.08GB ± 0%    0.00GB ± 0%  -100.00%  (p=0.008 n=5+5)

name                                     old allocs/op  new allocs/op  delta
ParseStatementNestedObjectsOrSets/1-16       69.0 ± 0%      77.0 ± 0%   +11.59%  (p=0.008 n=5+5)
ParseStatementNestedObjectsOrSets/5-16      1.32k ± 0%     0.24k ± 0%   -82.10%  (p=0.008 n=5+5)
ParseStatementNestedObjectsOrSets/10-16     42.8k ± 0%      0.4k ± 0%   -98.98%  (p=0.008 n=5+5)
ParseStatementNestedObjectsOrSets/15-16     1.37M ± 0%     0.00M ± 0%   -99.95%  (p=0.008 n=5+5)
ParseStatementNestedObjectsOrSets/20-16     43.9M ± 0%      0.0M ± 0%  -100.00%  (p=0.008 n=5+5)
```
for the super-un-optimal case, and this for the rest of the parser benchmarks:

```
name                                 old time/op    new time/op    delta
ParseModuleRulesBase/1-16              4.72µs ±12%    5.73µs ± 9%   +21.34%  (p=0.016 n=5+5)
ParseModuleRulesBase/10-16             31.6µs ± 4%    39.1µs ± 4%   +23.80%  (p=0.008 n=5+5)
ParseModuleRulesBase/100-16             297µs ± 5%     381µs ± 5%   +28.25%  (p=0.008 n=5+5)
ParseModuleRulesBase/1000-16           3.28ms ± 5%    3.87ms ± 7%   +17.73%  (p=0.008 n=5+5)
ParseStatementBasicCall-16             2.50µs ± 2%    3.09µs ± 5%   +23.79%  (p=0.008 n=5+5)
ParseStatementMixedJSON-16             75.8µs ± 2%   119.6µs ± 2%   +57.87%  (p=0.008 n=5+5)
ParseStatementSimpleArray/1-16         2.16µs ± 7%    2.59µs ± 2%   +19.84%  (p=0.008 n=5+5)
ParseStatementSimpleArray/10-16        7.35µs ±11%   10.82µs ±10%   +47.29%  (p=0.008 n=5+5)
ParseStatementSimpleArray/100-16       61.0µs ± 2%   102.1µs ±10%   +67.39%  (p=0.008 n=5+5)
ParseStatementSimpleArray/1000-16       585µs ± 5%     942µs ± 4%   +61.06%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x1-16     3.15µs ± 9%    3.82µs ± 4%   +21.26%  (p=0.008 n=5+5)
ParseStatementNestedObjects/5x1-16     8.14µs ± 4%   11.76µs ± 6%   +44.51%  (p=0.008 n=5+5)
ParseStatementNestedObjects/10x1-16    13.9µs ± 5%    21.4µs ± 4%   +53.60%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x5-16     7.90µs ± 4%   11.71µs ± 4%   +48.30%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x10-16    13.9µs ± 4%    21.3µs ± 5%   +53.57%  (p=0.008 n=5+5)
ParseStatementNestedObjects/5x5-16     5.02ms ± 2%    9.21ms ± 6%   +83.51%  (p=0.008 n=5+5)
ParseBasicABACModule-16                45.8µs ± 2%    60.7µs ± 3%   +32.48%  (p=0.008 n=5+5)

name                                 old alloc/op   new alloc/op   delta
ParseModuleRulesBase/1-16              3.97kB ± 0%    5.17kB ± 0%   +30.25%  (p=0.008 n=5+5)
ParseModuleRulesBase/10-16             26.0kB ± 0%    34.3kB ± 0%   +31.73%  (p=0.008 n=5+5)
ParseModuleRulesBase/100-16             252kB ± 0%     332kB ± 0%   +31.35%  (p=0.008 n=5+5)
ParseModuleRulesBase/1000-16           2.56MB ± 0%    3.46MB ± 0%   +35.21%  (p=0.008 n=5+5)
ParseStatementBasicCall-16             3.02kB ± 0%    3.94kB ± 0%   +30.18%  (p=0.008 n=5+5)
ParseStatementMixedJSON-16             51.9kB ± 0%   104.6kB ± 0%  +101.41%  (p=0.008 n=5+5)
ParseStatementSimpleArray/1-16         2.86kB ± 0%    3.77kB ± 0%   +31.93%  (p=0.008 n=5+5)
ParseStatementSimpleArray/10-16        5.86kB ± 0%    9.95kB ± 0%   +69.98%  (p=0.016 n=5+4)
ParseStatementSimpleArray/100-16       38.4kB ± 0%    78.0kB ± 0%  +103.32%  (p=0.008 n=5+5)
ParseStatementSimpleArray/1000-16       390kB ± 0%     845kB ± 0%  +116.76%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x1-16     3.66kB ± 0%    4.86kB ± 0%   +32.82%  (p=0.008 n=5+5)
ParseStatementNestedObjects/5x1-16     6.44kB ± 0%   10.54kB ± 0%   +63.63%  (p=0.008 n=5+5)
ParseStatementNestedObjects/10x1-16    10.0kB ± 0%    18.3kB ± 0%   +82.44%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x5-16     8.51kB ± 0%   12.61kB ± 0%   +48.14%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x10-16    14.6kB ± 0%    22.8kB ± 0%   +56.69%  (p=0.008 n=5+5)
ParseStatementNestedObjects/5x5-16     3.42MB ± 0%    6.94MB ± 0%  +103.21%  (p=0.008 n=5+5)
ParseBasicABACModule-16                35.1kB ± 0%    48.4kB ± 0%   +37.93%  (p=0.008 n=5+5)

name                                 old allocs/op  new allocs/op  delta
ParseModuleRulesBase/1-16                80.0 ± 0%      91.0 ± 0%   +13.75%  (p=0.008 n=5+5)
ParseModuleRulesBase/10-16                600 ± 0%       667 ± 0%   +11.17%  (p=0.008 n=5+5)
ParseModuleRulesBase/100-16             5.83k ± 0%     6.45k ± 0%   +10.63%  (p=0.008 n=5+5)
ParseModuleRulesBase/1000-16            58.0k ± 0%     64.1k ± 0%   +10.41%  (p=0.016 n=4+5)
ParseStatementBasicCall-16               47.0 ± 0%      55.0 ± 0%   +17.02%  (p=0.008 n=5+5)
ParseStatementMixedJSON-16              1.02k ± 0%     1.37k ± 0%   +34.51%  (p=0.008 n=5+5)
ParseStatementSimpleArray/1-16           36.0 ± 0%      44.0 ± 0%   +22.22%  (p=0.008 n=5+5)
ParseStatementSimpleArray/10-16           112 ± 0%       148 ± 0%   +32.14%  (p=0.008 n=5+5)
ParseStatementSimpleArray/100-16          837 ± 0%      1151 ± 0%   +37.51%  (p=0.008 n=5+5)
ParseStatementSimpleArray/1000-16       8.05k ± 0%    11.08k ± 0%   +37.71%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x1-16       49.0 ± 0%      60.0 ± 0%   +22.45%  (p=0.008 n=5+5)
ParseStatementNestedObjects/5x1-16        121 ± 0%       157 ± 0%   +29.75%  (p=0.008 n=5+5)
ParseStatementNestedObjects/10x1-16       207 ± 0%       274 ± 0%   +32.37%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x5-16        125 ± 0%       161 ± 0%   +28.80%  (p=0.008 n=5+5)
ParseStatementNestedObjects/1x10-16       220 ± 0%       287 ± 0%   +30.45%  (p=0.008 n=5+5)
ParseStatementNestedObjects/5x5-16      71.1k ± 0%     94.8k ± 0%   +33.25%  (p=0.008 n=5+5)
ParseBasicABACModule-16                   841 ± 0%       937 ± 0%   +11.41%  (p=0.008 n=5+5)
```

...but there's one more thing I'd like to try here....